Add general configuration. Move the configuration related to the storage to the storage implementation
