+++ caching-service/src/main/java/org/zowe/apiml/caching/service/vsam/VsamRecord.java
@@ -13,8 +13,8 @@ package org.zowe.apiml.caching.service.vsam;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.commons.lang3.StringUtils;
-import org.zowe.apiml.caching.config.VsamConfig;
 import org.zowe.apiml.caching.model.KeyValue;
+import org.zowe.apiml.caching.service.vsam.config.VsamConfig;
 
 import java.io.UnsupportedEncodingException;
