+++ caching-service/src/main/java/org/zowe/apiml/caching/service/vsam/VsamStorage.java
@@ -11,11 +11,11 @@ package org.zowe.apiml.caching.service.vsam;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.retry.annotation.Retryable;
-import org.zowe.apiml.caching.config.VsamConfig;
 import org.zowe.apiml.caching.model.KeyValue;
 import org.zowe.apiml.caching.service.Messages;
 import org.zowe.apiml.caching.service.Storage;
 import org.zowe.apiml.caching.service.StorageException;
+import org.zowe.apiml.caching.service.vsam.config.VsamConfig;
 import org.zowe.apiml.util.ObjectUtil;
 
 import java.util.*;
