+++ caching-service/src/test/java/org/zowe/apiml/caching/service/vsam/VsamRecordTest.java
@@ -12,8 +12,8 @@ package org.zowe.apiml.caching.service.vsam;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.zowe.apiml.caching.config.VsamConfig;
 import org.zowe.apiml.caching.model.KeyValue;
+import org.zowe.apiml.caching.service.vsam.config.VsamConfig;
 import org.zowe.apiml.zfile.ZFileConstants;
 
 import java.io.UnsupportedEncodingException;
